Add imageslide file
